A number of improvements and restructuring of files.
