+++ contentctl/actions/validate.py
@@ -1,14 +1,14 @@
 import pathlib
 
-from contentctl.input.director import Director, DirectorOutputDto
-from contentctl.objects.config import validate
 from contentctl.enrichments.attack_enrichment import AttackEnrichment
 from contentctl.enrichments.cve_enrichment import CveEnrichment
-from contentctl.objects.atomic import AtomicEnrichment
-from contentctl.objects.lookup import FileBackedLookup
-from contentctl.helper.utils import Utils
-from contentctl.objects.data_source import DataSource
 from contentctl.helper.splunk_app import SplunkApp
+from contentctl.helper.utils import Utils
+from contentctl.input.director import Director, DirectorOutputDto
+from contentctl.objects.atomic import AtomicEnrichment
+from contentctl.objects.config import validate
+from contentctl.objects.data_source import DataSource
+from contentctl.objects.lookup import FileBackedLookup
 
 
 class Validate:
@@ -17,6 +17,7 @@ class Validate:
             AtomicEnrichment.getAtomicEnrichment(input_dto),
             AttackEnrichment.getAttackEnrichment(input_dto),
             CveEnrichment.getCveEnrichment(input_dto),
+            [],
             [],
             [],
             [],
