+++ contentctl/objects/abstract_security_content_objects/security_content_object_abstract.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Any, Self
 if TYPE_CHECKING:
     from contentctl.input.director import DirectorOutputDto
     from contentctl.objects.deployment import Deployment
-    from contentctl.objects.config import CustomApp, Config_Base
+    from contentctl.objects.config import Config_Base
 
 import abc
 import datetime
@@ -13,7 +13,6 @@ import pathlib
 import pprint
 import uuid
 from abc import abstractmethod
-from csv import DictWriter
 from functools import cached_property
 from typing import List, Optional, Tuple, Type, Union
 
@@ -37,30 +36,29 @@ from contentctl.objects.constants import (
     DEPRECATED_TEMPLATE,
     EXPERIMENTAL_TEMPLATE,
 )
-
-# from contentctl.objects.baseline import Baseline
-# from contentctl.objects.detection import Detection
-from contentctl.objects.enums import AnalyticsType, DetectionStatus, SecurityContentType
+from contentctl.objects.enums import AnalyticsType, DetectionStatus
 
 NO_FILE_NAME = "NO_FILE_NAME"
 
 
 class DeprecationInfo(BaseModel):
     deprecated_content: SecurityContentObject_Abstract
-    deprecation_date: datetime.date = Field(
+    deprecated_date: datetime.date = Field(
         ...,
         description="On what expected date will this content be deprecated? "
         "If an app is built after this date and attempts to write out this content, an exception will be generated.",
     )
-    deprecation_version: str = Field(
+
+    deprecated_in_version: str = Field(
         ...,
         description="In which version of the app was this content deprecated? "
         "If an app is built on or after this version and contains this content, an exception will be generated.",
     )
+
     reason: str = Field(
         ...,
         description="An explanation of why this content was deprecated.",
-        min_length=10,
+        min_length=0,
     )
     replacement_content: list[SecurityContentObject_Abstract] = Field(
         [],
@@ -85,7 +83,7 @@ class DeprecationInfo(BaseModel):
             else:
                 raise Exception(
                     f"Content named '{self.deprecated_content.name}' "
-                    f"was marked for deprecation in version [{self.deprecation_version}]. "
+                    f"was marked for deprecation in version [{self.deprecated_in_version}]. "
                     f"However, this content is STILL PRESENT in the current build [{cfg.app.version}]. "
                     f"This content should be moved to the folder '{self.deprecated_content.file_path}' into '{cfg.deprecated_content_path}'"
                 )
@@ -93,9 +91,9 @@ class DeprecationInfo(BaseModel):
             if self.hasContentBeenRemoved(cfg):
                 raise Exception(
                     f"Content named '{self.deprecated_content.name}' "
-                    f"was marked for deprecation in version [{self.deprecation_version}]. "
-                    f"However, this content is HAS BEEN REMOVED in the current build [{cfg.app.version}]. "
-                    f"This content should be moved out of the folder '{cfg.deprecated_content_path}'"
+                    f"was marked for deprecation in version [{self.deprecated_in_version}]. "
+                    f"However, this content is HAS BEEN REMOVED EARLY in the current build [{cfg.app.version}]. "
+                    f"This content should be moved out of the folder '{cfg.deprecated_content_path} and into its appropriate content folder.'"
                 )
             else:
                 print(
@@ -120,6 +118,7 @@ class DeprecationInfo(BaseModel):
                 f"has been moved into the folder {cfg.deprecated_content_path}. "
                 "The content is not backed by a file (content.file_path was 'None')"
             )
+
         return self.deprecated_content.file_path.resolve().is_relative_to(
             cfg.deprecated_content_path.resolve()
         )
@@ -138,10 +137,10 @@ class DeprecationInfo(BaseModel):
             bool: True or False, based on whether something should be removed from the app
         """
         try:
-            deprecation_version = Version(self.deprecation_version)
-        except Exception:
+            deprecation_version = Version(self.deprecated_in_version)
+        except Exception as e:
             raise Exception(
-                f"Unable to parse deprecation_version info for {self.deprecated_content.name} into a valid Semantic Version: [{self.deprecation_version}]"
+                f"Unable to parse deprecation_version info for {self.deprecated_content.name} into a valid Semantic Version: [{self.deprecated_in_version}]: {e}"
             )
         try:
             current_app_version = Version(cfg.app.version)
@@ -149,9 +148,20 @@ class DeprecationInfo(BaseModel):
             raise Exception(
                 f"Unable to parse deprecation_version info for the app {cfg.app.title} into a valid Semantic Version: [{cfg.app.version}]"
             )
+        print(f"Dep: {deprecation_version}, App: {current_app_version}")
         if deprecation_version <= current_app_version:
             return True
         return False
+
+    @field_validator("replacement_content", mode="before")
+    @classmethod
+    def mapReplacementContent(
+        cls, v: list[str], info: ValidationInfo
+    ) -> list[SecurityContentObject_Abstract]:
+        director: DirectorOutputDto = info.context.get("output_dto", None)
+        return SecurityContentObject_Abstract.mapNamesToSecurityContentObjects(
+            v, director
+        )
 
 
 class DeprecationDocumentationFile(BaseModel):
@@ -188,14 +198,18 @@ class DeprecationDocumentationFile(BaseModel):
                 )
             try:
                 elem["deprecated_content"] = (
-                    contentClass.mapNamesToSecurityContentObjects([b], director)
+                    contentClass.mapNamesToSecurityContentObjects([name], director)[0]
                 )
             except Exception:
                 try:
+                    from contentctl.objects.deprecated_security_content_object import (
+                        DeprecatedSecurityContentObject,
+                    )
+
                     elem["deprecated_content"] = (
-                        SecurityContentObject_Abstract.mapNamesToSecurityContentObjects(
-                            [elem], director
-                        )
+                        DeprecatedSecurityContentObject.mapNamesToSecurityContentObjects(
+                            [name], director
+                        )[0]
                     )
                 except Exception:
                     raise ValueError(
@@ -576,109 +590,6 @@ class SecurityContentObject_Abstract(BaseModel, abc.ABC):
         return id(self)
 
 
-class DeprecatedSecurityContentObject(SecurityContentObject_Abstract):
-    # We MUST allow extra fields here because the python definitions of the underlying
-    # objects can change. We do not want to throw pasing errors on any of these, but we will
-    # only expose fields that are defined in the SecurityContentObject definiton directly
-    model_config = ConfigDict(validate_default=True, extra="ignore")
-
-    @field_validator("deprecation_info")
-    def ensure_deprecation_info_is_not_none(cls, deprecation_info: Any) -> Any:
-        if deprecation_info is None:
-            raise ValueError(
-                "DeprecatedSecurityObject does not define a valid deprecation_info object."
-            )
-        return deprecation_info
-
-    @computed_field
-    @cached_property
-    def migration_guide(self) -> HttpUrl:
-        """
-        A link to the research site containing a migration guide for the content
-
-        :returns: URL to the research site
-        :rtype: HTTPUrl
-        """
-        # this is split up so that we can explicilty ignore the warning on constructing
-        # the HttpUrl but catch other type issues
-        link = f"https://research.splunk.com/migration_guide/{self.id}"
-
-        return HttpUrl(url=link)  # type: ignore
-
-    @staticmethod
-    def writeDeprecationCSV(
-        deprecated_content: list[DeprecatedSecurityContentObject],
-        app: CustomApp,
-        output_file: pathlib.Path,
-    ):
-        deprecation_rows: list[dict[str, str]] = []
-        for content in deprecated_content:
-            if content.deprecation_info is None:
-                raise Exception(
-                    f"Cannot compute deprecation info for {content.name} - object has no deprecation info"
-                )
-            content.generateDeprecationLookupRow(content, app)
-
-        with open(output_file, "w") as deprecation_csv_file:
-            deprecation_csv_writer = DictWriter(
-                deprecation_csv_file,
-                fieldnames=[
-                    "Name",
-                    "ID",
-                    "Content Type",
-                    "Deprecation Date",
-                    "Deprecation Version",
-                    "Reason",
-                    "Migration Guide",
-                    "Replacement Content",
-                ],
-            )
-            deprecation_csv_writer.writeheader()
-            deprecation_csv_writer.writerows(deprecation_rows)
-
-    def generateDeprecationLookupRow(self, app: CustomApp) -> dict[str, str]:
-        """
-        This function exists as a bit of a shim because pieces of content are identified in ESCU by
-        their Stanza name, which could be different than
-        """
-        if self.deprecation_info is None:
-            raise ValueError(
-                f"DeprecatedSecurityContentObject 'deprecation_info' field was None for '{self.name}' in file '{self.file_path}'"
-            )
-
-        if self.deprecation_info.content_type == SecurityContentType.detection:
-            from contentctl.objects.detection import Detection
-
-            content_name = Detection.static_get_conf_stanza_name(self.name, app)
-        elif self.deprecation_info.content_type == SecurityContentType.story:
-            content_name = self.name
-        elif self.deprecation_info.content_type == SecurityContentType.baseline:
-            from contentctl.objects.baseline import Baseline
-
-            content_name = Baseline.static_get_conf_stanza_name(self.name, app)
-
-        else:
-            raise Exception(
-                f"{self.deprecation_info.content_type} deprecation is not supported at this time."
-            )
-
-        return {
-            "Name": content_name,
-            "ID": str(self.id),
-            "Content Type": self.deprecation_info.content_type.name,
-            "Deprecation Date": str(self.deprecation_info.deprecation_date),
-            "Deprecation Version": str(self.deprecation_info.deprecation_version),
-            "Reason": "Give a unique reason in the model here",
-            "Migration Guide": str(self.migration_guide),
-            "Replacement Content": "\n".join(
-                [
-                    str(content.researchSiteLink)
-                    for content in self.deprecation_info.replacement_content
-                ]
-            ),
-        }
-
-
 # class DeprecatedSecurityContentObject(SecurityContentObject_Abstract):
 #     # We MUST allow extra fields here because the python definitions of the underlying
 #     # objects can change. We do not want to throw pasing errors on any of these, but we will
@@ -799,6 +710,8 @@ class DeprecatedSecurityContentObject(SecurityContentObject_Abstract):
 #             ),
 #         }
 #             ),
+#         }
+#         }
 #         }
 #         }
 #         }
